Fixed: Allocation size truncation in ALLOC macros on 64-bit systems
Removed unsafe (int) casts from sizeof() in ALLOC and ALLEXPR macros.
On 64-bit systems, sizeof() returns size_t (64-bit) and casting to int
could theoretically truncate, though struct sizes are small in practice.

+++ src/defines.h
@@ -6,8 +6,8 @@
 
 #define M(x) (1<<x)	/* Mask (x) returns 2^x */
 
-#define ALLOC(x)	(struct x *) ckalloc((int)sizeof(struct x))
-#define ALLEXPR		(expptr) ckalloc((int)sizeof(union Expression) )
+#define ALLOC(x)	(struct x *) ckalloc(sizeof(struct x))
+#define ALLEXPR		(expptr) ckalloc(sizeof(union Expression) )
 typedef int *ptr;
 typedef char *charptr;
 typedef FILE *FILEP;
